Sync with my 2.1.0.1 version.
Preference to start CLI.  This will start the Mininet CLI in the window you started MiniEdit.  Note that there is a warning about this feature.  You will need to remember to .quit. the CLI as well whenever you .Stop. your mininet if you enable this feature.  By default, CLI option is off, but it can be turned on in the Preferences.
Made some improvements to the Export code to also export all your Preferences.
If you decide to upgrade your Open vSwitch in the VM to Open vSwitch 1.10 or higher, you now have the option in MiniEdit of choosing which versions of OpenFlow to enable.  OVS 1.10 added support for choosing 1.0, 1.2 and 1.3.
A new feature in Mininet 2.1.0 is experimental support of Indigo Virtual Switch.  I have added a Preference so you can choose to use Indigo Virtual Switch or Open vSwitch.