Update LincSwitch in node.py so that multiple instances can be run
Mininet requires that LINC-Switch can be started in multiple instances
on a single machine. Each instance has to be run from a separate release.
Thus we generate a separate release for each instance and configuration
files for them. We also take care of removing generated release when
stopping Mininet.
(Motivated by FlowForwarding/LINC-Switch#240).