Add VPP as a possible switch
VPP stands for Vector Packet Processing, and is a  is a rapid packet processing
development platform for highly performing network applications.
See the "what is VPP" page: https://wiki.fd.io/view/VPP/What_is_VPP%3F
    and the fd.io web page: https://fd.io

The integration is done using the python API, and presenting the vpp cli
to mininet via dpctl.

vpp import is left within the VPPSwitch class so that this does not
become a new dependency.