Added support for KDE Neon
KDE Neon is derived from Ubuntu LTS and so everything is on
a 2 year release cycle. However, Plasma on top is rolling release.
Because of the way Plasma tends to be developed,
rolling Plasma tends to not only be more stable,
but you also get all the nice features and bug fixes sooner.
So it's safe to treat KDE Neon as Ubuntu.